Update Selector definition
The consensus from the 12/14/16 meeting was to simplify this definition
to use plainer text. As well, Alistair asks to be careful about the use
of element vs. node.

Some simplified text suggestions:
"selectors are patterns used to select the element(s) you want to work
with." - Alistair
From CSS 3 Selectors, Selectors are patterns that match against elements
in a tree, and as such form one of several technologies that can be used
to select nodes in an XML document. Selectors have been optimized for
use with HTML and XML, and are designed to be usable in
performance-critical code.
From w3 Schools: In CSS, selectors are patterns used to select the
element(s) you want to style.